refactor(frontend): extract generic CustomFieldsInput component out of Case-Incident code
